+++ muxers/mplex/src/lib.rs
@@ -169,7 +169,7 @@ where
         self.io.lock().drop_stream(sub.id);
     }
 
-    fn close(&self, cx: &mut Context<'_>) -> Poll<Result<(), io::Error>> {
+    fn poll_close(&self, cx: &mut Context<'_>) -> Poll<Result<(), io::Error>> {
         self.io.lock().poll_close(cx)
     }